alms-btt/almsnew/backend/requirements.txt
fastapi==0.99.0
uvicorn==0.22.0
sqlalchemy==2.0.15
pydantic==1.10.9
python-dotenv==1.0.0
msal==1.22.0
numpy==1.24.3
pandas==2.0.2
pyodbc==4.0.39
redis==4.5.5
python-jose==3.3.0
passlib==1.7.4
email-validator==2.0.0
httpx==0.24.1
pytest==7.3.1
requests==2.31.0 